refactor: break out index.js into seperate files
