+++ Server/config/database.js
@@ -1,0 +1,14 @@
+const { PrismaClient } = require('@prisma/client');
+const { PrismaLibSQL } = require('@prisma/adapter-libsql');
+const { createClient } = require('@libsql/client');
+const { DATABASE_URL, DATABASE_AUTH_TOKEN } = require('./keys');
+
+const libsql = createClient({
+  url: DATABASE_URL,
+  authToken: DATABASE_AUTH_TOKEN,
+});
+
+const adapter = new PrismaLibSQL(libsql);
+const prisma = new PrismaClient({ adapter });
+
+module.exports = prisma;
