+++ Server/config/keys.js
@@ -1,0 +1,4 @@
+module.exports = {
+  DATABASE_URL: process.env.TURSO_DATABASE_URL,
+  DATABASE_AUTH_TOKEN: process.env.TURSO_AUTH_TOKEN,
+};
