+++ Server/routes/healthRoutes.js
@@ -1,0 +1,8 @@
+const express = require('express');
+const healthController = require('../controllers/healthController');
+
+const router = express.Router();
+
+router.get('/health', healthController.checkHealth);
+
+module.exports = router;
